Clean up base64 provided for responses.
If extra characters like newlines are included in base64 responses then
the response we send is garbled.  Since newline breaks are actually
relatively common for base64 this can be confusing.

This actually removes *all* non base64 characters from the string before
generating the response.

This also removes the documentation note about ensuring base64 is
provided on a single line since that will no longer be true.