Cleaned up stubIndex(); prevented most race conditions
There are still edge cases where state changes between the call
to stubIndex() and the call to insertAtIndex() (for proxyOnce)
or the call to first(.., index) (for proxyAlways). Except for
degenerate scenarios where a user is calling stub-changing APIs
in the middle of proxy recording, the semantics of proxyOnce
and proxyAlways will still be obeyed even if downstream proxy calls
return in unpredictable orders.
https://github.com/bbyars/mountebank/issues/279