Resolved locking conflicts adding imposters
Prior to this change, every stub addition was locking the header file
in addition to the actual imposters.add call. Moving the initial creation
of all stubs to imposters.add removes all file locking from the creation
of an imposter. With this change, nearly all the tests pass with the new
repository.
https://github.com/bbyars/mountebank/issues/279