Update docker cmd-line in Getting Started
The example imposters in the Getting Started page use ports 4545 and 5555 . The docker container is not exposing them but 8080. AFAIU 8080 is not used in anyway, and only the controller at 2525 needs to be additionally exposed.

The file `imposters.ejs` does not seem to exist in the container, but I believe it is not necessary for this example either.